refactor(notification): domain 내 sse 패키지 분리
