--- backend/foodtruck/src/main/java/com/connectruck/foodtruck/notification/domain/SseEventGroupType.java
@@ -1,6 +1,0 @@
-package com.connectruck.foodtruck.notification.domain;
-
-public enum SseEventGroupType {
-
-    OWNER_ORDER
-}
