+++ backend/foodtruck/src/main/java/com/connectruck/foodtruck/notification/domain/sse/RedisSseEventRepository.java
@@ -1,4 +1,4 @@
-package com.connectruck.foodtruck.notification.domain;
+package com.connectruck.foodtruck.notification.domain.sse;
 
 import java.util.List;
 import org.springframework.data.redis.core.RedisTemplate;
