+++ backend/foodtruck/src/main/java/com/connectruck/foodtruck/notification/domain/sse/SseEventGroup.java
@@ -1,4 +1,4 @@
-package com.connectruck.foodtruck.notification.domain;
+package com.connectruck.foodtruck.notification.domain.sse;
 
 import static lombok.AccessLevel.PROTECTED;
