+++ backend/foodtruck/src/main/java/com/connectruck/foodtruck/notification/domain/sse/SseEventGroupType.java
@@ -1,0 +1,6 @@
+package com.connectruck.foodtruck.notification.domain.sse;
+
+public enum SseEventGroupType {
+
+    OWNER_ORDER
+}
